Use factory methods on Request struct
This way, the logic for using the correct way of initializing the
params field is encapsulated within the Request struct implementation.
As the default ctor ( Request { version: ... etc } ) is disabled by
default if the fields are private, the only way of creating an
instance of Request is through one of the factory methods. And those
are now unit-tested :)