Print logs of docker containers if tests fail
We introduce an environment variable KEEP_CONTAINERS_AFTER_TEST. If
this one is true, the `Drop` implementation of the nodes will `stop`
the containers instead of `rm`-ing them.

In the .travis.yml file, we check for the TRAVIS_TEST_RESULT variable
and detect whether the tests failed. If they failed, we print all the
logs of all the containers to stdout.

To facilitate local development, the fallback implementation of
KEEP_CONTAINERS_AFTER_TEST is false, so that we don't accumulate
containers locally.