Introduce FetchTransactionStream
This is an extension trait for LedgerQueryServiceApiClient.

The extension trait configures a given `Ticker` stream to repeatedly
poll the LedgerQueryService for results and uniquely emits them as
they come in.

This resolves multiple things:

- It gets rid of the whole futures_ext abstraction that I built
before. We actuallt don't need that if we build extension traits on
top of existing traits.
- Through that, it fixes #311.
- The new design also allows for a better testing.