Fix race condition with unexpected response
What was the bug?

The problem with the old implementation was that the lock on the
response_source was acquired AFTER the frame was sent. This made it
possible that with unlucky timing, the response was received before
this part of the code acquired the lock. As a result, the frame was
discarded as `Unexpected frame` because it was not yet registered as
by the code in the client.

How does this change fix this bug?

The important bit is that we have to construct the response-future
before we send the frame. We do this by locking the ResponseSource
right away, followed by reserving the id of this frame. Given that,
we can construct the response-future, and compose that together with
the future that sends the request.