Remove unnecessary From implementation
- Remove trait bound `From<Self::Address>` on associated type `Identity`
of trait `Ledger` because it is not needed. This has the side-effect
of allowing the removal of a fallible from implementation which is
something that `clippy` warned about and had to be ignored. Next time
`TryFrom` may be able to be used as it's coming to stable soon.
- Delete commented out test because the function that it could be
testing doesn't exist anymore.