panic! instead of passing error to caller
There is not really a point in passing on the error that can happen
during serialization of a request since it is something that either
ALWAYS fails or never. Meaning, you will catch this error at
development/testing time and if it doesn't occur there, it will not
happen in production.