Refactor actions to embed validity information
Instead of a generic invalid_until field, we embed the mediantime
timestamp for Bitcoin and the block timestamp for Ethereum directly
into the actions.

This allows us to implement customized "wait" strategies depending
on the ledger.

- For Bitcoin, we check the mediantime from the
getblockchaininfo RPC call and wait until it passes the locktime
threshold.

- For Ethereum, it is enough to wait until our local systemtime
passes the given timestamp.