Add hash_function parameter to rfc003 'protocol' header
Currently we do not have a hash_function parameter for the rfc003 SWAP
protocol header.  This is an omission because the RFC-003 describes the
hash_function parameter as mandatory for the protocol.

We do not currently want users of comit-rs to be able to configure the
hash_function, we should therefore not expose this parameter via the
REST API.

Add hash_function as a parameter to the BAM API 'protocol' header for
rfc003 swaps.  Add a unit test to check conversion between SwapProtocol
objects and BAM::Header objects.