Use `git diff --name-only`
We can use `git diff --name-only` to make this script nicer.  We need to
run the command twice, with and without `--cached` to catch all work
flows.  We can throw this in a loop and use global variables to store
the result.  Make decision on which format checker to run based on the
stored results.

Suggested by: Thomas Eizinger