Rename PendingIncomingRequest -> PendingOutgoingRequest
This enum is currently named incorrectly.  The event is an 'in' event,
i.e., the event occurs in the this node (as opposed to peer node).  The
event we are pending is then an outgoing request.  This is also shown by
the fact that the request member type is 'PendingOutgoingRequest'.