Move print statements to single method
Currently we print to stdout/stderr from three different methods.  Its
unfortunate that we have to print to stdout/stderr at all (no use the logger),
if we put all the print statements in a single method it is easier to see what
output File should be giving us at run time.