Refactor to mirror cnd `read` in btsieve
We have very similar code in btsieve and cnd for parsing config files.  It is
quicker to grok the code if functions that are the same in both are written the
same.

Refactor the `read` function in btsieve to be identical to the same function in
cnd.  This patch is refactor only, makes no logic changes.