Move command line options struct to separate module
Currently we declare the command line options struct within
`load_settings.rs`, since there is only a single command line option
this is ok but it is cleaner if the command line options are separate
from any particular option.  This is in line with how we do it in cnd.

Move command line options struct to separate module mirroring the
command line option code in cnd.