Make `directories` usage uniform
We use the `directories` crate in both btsieve and cnd.  We should use it in a
uniform manner.  Wrap the usage of `directories` in a function call to improve
maintainability.  Do the same in both btsieve and cnd.  While we are at it use
combinators to handle the error return code path instead of using `if let`, its
cleaner to have the error code indented as a combinator.