Use Sqlite backed MetadataStore
Add and use an Sqlite database for storing `Metadata` on disk.

Add configuration option to configure the path to the Sqlite
database file.  With this applied cnd can be configured with

`cnd.toml`:

        [database]
        sqlite = "/path/to/sqlite"

This config option is optional, without it we use the default project
data directory:

        Linux: /home/<user>/.local/share/comit/cnd.sqlite
        OSX: /Users/<user>/Library/Application Support/comit/cnd.sqlite

Add database migrations for use with diesel to the directory
`cnd/migrations/`.  Currently this directory must be copied to the
project data directory (same as sqlite config file above).

Migrations are applied in code but have been created manually, the
migrations will need to be created in code in the future.

With this set patch applied e2e tests use a temporary Sqlite db file for
each test.