Disable SqliteMetadataStore
Sqlite database implementation is incomplete; currently we only store
the `Metadata` in the database and not the `StateStore`.  This means cnd
returns a 500 error for the swaps/ endpoint.

Disable usage of the `SqliteMetadataStore`.  This is done as a separate
commit on top of the Sqlite stuff so we can add dormant code to the
codebase without introducing regressions.  This could have been done in
code but the added complexity would be a waste of time to implement.