Remove ethereum query transaction boolean option
Like we did in cnd just now we can use a boolean directly instead of a
boolean wrapped in an Option.  The Option adds no extra information.

This patch replaces `Option<false>` with false, this is in contradiction
to what is stated above.  `Option<false>` implies there is a difference
between `Option<false>` and `None`, the patch author claims this is an
incorrect implication.  We have good unit test coverage for the
`matches()` function.