Replace String associated data in Error variants
Currently database errors are define in the metadata_store module, they
should be in the database module.  We can then just have a single Error
enum variant in metadata_store with the Sqlite error as its associated
type.  Secondly, the error variants all use String as their associated
type, this is lazy, we can do better than this.

- Move all the Sqlite specific Error enum variants to the database
module.

- Create a single Error enum variant in metadata_store that has an
Sqlite error as its associated data.

- Remove the use of String types from associate data for the custom
Error used by the db module.  If there is another type of error
available use that as the associated type.

- Use `use crate::metadata_store as ms` to make the code less verbose.