Add COMIT message event timestamps
Add timestamps for all events that are make up the COMIT message
exchange for rcf003.  We store the timestamps in the `SwapCommunication`
structure for Bob and Alice.  This means that, in line with how the rest
of the code uses this struct, each party has a timestamp for the event
as it occurred for them.

The following events are timestamped:

- swap request sent/received
- swap decline sent/received
- swap accept sent/received