Mirror "network" functionality from Bitcoin for Ethereum
Currently, we are _not_ validating whether the URL we are given
for a Bitcoin Blocksource actually points to the network we are
also passing.
For Ethereum though, we used to fetch it from the node we got pointed
at. However, now that we are merging btsieve into cnd, this prevents
cnd from starting up without having ledgers configured because
it would try and connect to the default url of Ethereum.

This patch removes this functionality and passes the network in
explicitely.