Make URLs for Blocksources type-safe
The previous approach created problems with incorrectly generated
URLs like: http://localhost:8080//rest/...

(Note the double slash)

Unfortunately, we have to build URLs at runtime for the blocksources
to work, hence we cannot avoid a possible runtime panic here. To
mitigate this issue, we extensively test the functions that build these
URLs to make sure that, regardless of the input, they never panic.