Use BlockId instead of String
When doing property based testing on a hash data field (of type
`String`) quickcheck generates input that causes a panic.  The generated
input is not possible during normal program flow.  To inhibit quickcheck
from generating input that is not useful to us we can use a specific
type instead of a string.  We already have code in tree doing the same
sort of thing, we can do the same here.

Use `bitcoin_support::BlockId` instead of `String`.

Resolves: #1506