Move secret seed into a separate file
Move the secret seed out of the cnd config file and into its own
separate file.

Use the pem format for the file:

	-----BEGIN SECRET SEED-----
	mbKANv2qKGmNVg1qtquj6Hx1pFPelpqOfE2JaJJAMEg1FlFhNRNlFlE=
	-----END SECRET SEED-----

The payload is 32 bytes encoded as base64.

Add a config option (and remove secret_seed option)

	[comit]
	secret_seed_file = "/tmp/secret_seed.pem"

Add a command line optio

	cnd --secret-seed-file

Order of precedence is as usual:

1. command line option
2. config file
3. default location