Unify HTTP API dependency trait bounds
The new `Connector` struct implements a bunch of trait bounds.  Instead of
explicitly stating the ones we want we can provide a new trait that collects all
of them together.  The compiler will produce the same code after this patch
since there is only one object in the code base that satisfies these trait
bounds.

Add a trait `Connect` and use it throughout the `http_api` module.