Remove Dependencies
Currently the `ComitNode` and the `http_api::Dependencies` struct both
embed a `Dependencies` struct.  This implies they are connected more
than they actually are.

We just removed the need for the spawner to access the swap seed.  There
are now on barriers to removing this 'false' connection created by the
`Dependencies` struct, we can embed all the `Dependencies`s members
directly into the `ComitNode` and the `http_api::Dependencies` struct.