Remove possibility for in-memory DB
It doesn't make much sense to have it because of the way our
DB code is setup.

In-memory DBs of SQLite only live as-long as the connection is open.
This means, we drop the DB just after we ran the migrations. In
order to properly support in-memory DBs, we'd have to rework the
way our DB code handles the connections as we would need to keep
them around. In-memory DBs would only be useful for testing and
we don't need them for now, hence it is not worth supporting them
properly and certainly not in-properly.