Add test that verifies that cnd does not (yet) persist any data
We achieve this by adding a `restart` function on the new Actor class.
To avoid code duplication between the old and the new test framework,
we extract the common code into a CndInstance class that is responsible
for starting and stopping instances of cnd.

The new Actor class will now spawn their own instance of cnd. This
brings us one step closer to tests that are self-contained.