Use Scribe instead of Debug
We would like to reduce the amount of noise in log output.  Use new
`Scribe` trait for generating log output instead of `Debug` so we can
control the output.  So as not to be a chore supply a default
implementation of `Scribe` that uses `Debug` - then we can just
implement `Scribe` when the logs are too noisy.