Add two types `RootSeed` and `SwapSeed`
Currently we have a `Seed` type that is used as the root seed to derive
per swap seeds which are then used to derive identities (refund/redeem).
Using the same type for all of these seeds means the compiler cannot
give us type safety.  Since these two things are separate we can use
separate types to allow the compiler to differentiate the types and
provide safety.

Add `RootSeed` for the seed read in from the PEM file and derive a new
`SwapSeed` to use for identity derivation.