Migrate init_accepted_swap to tokio 0.2
Tokio 0.2 no longer provides an `Executor` trait. Instead, we
use the `tokio::task::spawn` method to spawn an async task in
the background.
This is not as explicit as calling `.spawn()` on dependencies,
but as long as we only do this from application code (read cnd)
and not from potential library code, this should be fine.