Automatically stage rust files staged by cargo fmt
Our makefile only formats files that are staged. Formatting them
unstages them and hence, they need to be restaged again before they
can be committed.
A recent version of rustfmt added the option `--files-with-diff`
which outputs the files that were affected by the formatting.
We can use that information to automatically restage those files
and hence save a manual step for the developer. A usual workflow
would now be:

- write code
- stage the files ready for a commit
- make format
- commit