Use custom Infallible
`std::convert::Infallible` is meant to be used as error type (according
to the docs), seeing it in the `Ok` variant of `Result` is confusing.
We are using `Infallible` to make explicit that no `Ok` result is ever
returned by this function.  Using the std infallible violates the
principle of least surprise.  We can better achieve this behaviour by
adding a custom `Infallible` type and using it.