Move 'look back in time' outside of loop
We only look back in time on the first iteration through the loop, the
code is easier to reason about if this is done outside of the loop.
Doing so adds some code duplication but this is, IMO, better than the
added complexity.  Simple code is easier to read, easier to reason
about, and easier to find bugs in.