Check cnd is up before making action assertions
Currently during certain actions (fund, refund etc) we make assertions
on the other actor i.e., for bob we assert alice has seen the action.
This does not work if we are doing tests with one actors cnd node down.

Add if guards to check an actor's cnd is up before making assertions on
it.