Don't generate blocks as part of the test
Previously, we generated 101 blocks _for each test_ to ensure we
can spend the coinbase transactiosn and hence can send money to
the wallet address.

We've also been using the `generate` RPC call which will be phased
out in the future. It is recommended to use `generateToAddress` instead.

We should strictly avoid generating blocks from two different sources
because it seems that it creates race conditions inside bitcoin of
which "generate" RPC call wins and is actually able to produce a single
chain. We generate a new block every second, hence it is likely that
trying to generate 101 from a 2nd source causes problems.

To solve this, we move the ensureFunding call to the beginning of the
test harness and simply fail inside the wallet if for some reason the
wallet is used from a context that doesn't have the first 101 blocks
generated.