Index latest blocks from 0
Recent change indexed 'latest blocks' from 1, this was incorrect.  We should index
the latest blocks vector from 0 and ensure that a previous block is available in
'all blocks' vector.

For completeness; we set the start time of swap to the time in block 2, this is
just a convenience and works with `predates` function (which uses `<`).