Unroll some implementations generic over ledger::Bitcoin
If we want to split a comit crate out of cnd, we cannot abstract
over a trait that is not defined in our crate.

This patch removes all the usages of the ledger::Bitcoin that are
used in code that would be outside of the comit crate for abstracting
over any bitcoin network.