Unroll Bitcoin trait
Currently we are using a trait bound bound to `Bitcoin`.  Due to the
fact that rustc does not use trait bounds when differentiating impl
blocks this is currently buggy, the impl blocks are actually blanket
implementations.  This currently works because we only use the
implementations for the bitcoin side, as soon as we try to add the
ethereum side this breaks.

Attempt to unroll the `Bitcoin` trait and use concrete network types.
This all goes ok until we get to the database layer.  Here we cannot
unroll it because our current db implementation does not allow different
types to be in the same column (by design).  With concrete network types
all the database tables break - we would need to 'unroll' the tables
also i.e., triple the tables, one for each network.