Re-write the macaroon location logic
Re-write the logic for finding the lnd macaroon file.  We now expose two
functions to locate the macaroon, one library function and a default
caller on top of that

- locate_macaroon()       \\ takes a vector of directories to look in
- locate_macaroon_in_default_places()

The second function looks in the following places in an OS agnostic
manner.  For Linux this would be:

1. ~/.cache/comit/
2. ~/.cache/comit/lnd
3. ~/.lnd/
4. ~/.lnd/data/chain/bitcoin/regtest/

The macaroon file name used is `readonly.macaroon`.