Improve lnd helper functions
Generalise the macaroon location logic and add functions for locating
the lnd tls cert and key.

Note, this also makes the assumption that all lnd files, if in the cnd
data directory will be in a sub-directory called `lnd`.  This is done,
as commented in the code, so we can use generic terms like `tls.cert` as
are used by lnd.