Don't manually map the given HttpLedger to a Bitcoin ledger
Prior to this patch, we manually mapped the enum variant of HttpLedger
to an instance of the Bitcoin ledger in post_swap.rs

This is error prone and can lead to bugs where we are given one ledger
in the enum variant but instantiate another one for the actual swap.

With this patch we make the Bitcoin ledger a field of the enum variant.
This way, we can just pass on the instance in post_swap.rs.
The mapping is now defined in a single place where bugs are theoretically
still possible but easier to spot.