Run make format with the newly updated toolchain
Recently we updated the toolchain to 1.41.1, and also the nightly
toolchain to nightly-2020-01-15. CI is configured to run a different
version of rustfmt so this got past CI. However, this upgrade introduce
a bunch of new formatting to our codebase.

Run make format
Configure CI to use the same nightly toolchain as the Makefile