Split swap state into multiple stores
COMIT's core provides updates on the state of the ledgers during a
swap. These updates used to be applied directly on a struct which held
all the information pertaining to the swap: offer, communication
state, ledger state and error state. Storing everything together is
convenient but creates dependencies which cause pain.

The code can be simplified by splitting up this state into multiple
stores. Primarily, a generic ledger store (instantiated twice, for
alpha and beta) can be used to fulfil the core of COMIT's principal
task.

The communication state also needs to be saved and updated given the
current communication protocol. Since this protocol is going away
soon, it is not that appealing to invest a lot of effort into
improving the design, so it is simply stored in a separate `HashMap`.
This is not ideal because once the swap is accepted the information
can also be obtained from the database layer, meaning that we are
storing redundant information. Since it is easier to add another
`HashMap` than adding APIs to the database to store and load proposed
swaps, and this communication protocol is on the way out, incurring in
this technical debt is justified.

In terms of the error state, it turns out that we were never setting
the state of the swap to failed if watching a particular ledger had
resulted in an error. Since this was the apparent intention when
introducing these APIs, the functionality has been kept. It has also
been implemented using a `HashMap`, but a better design may come in
the future.