Only start one instance of Parity, Bitcoind and one lnd per actor
By utilizing file-locks on directories, we can ensure that any number
of parallel test suites only ever get to run one instance of
the ledger. The remaining test suites will take the config and
pass that to the tests.

We store the PID of each ledger in a file inside the lock directory,
this also includes a sub-process that we spawn for the Bitcoin miner.
This is important, because same as with the other nodes, this process
must keep going until the last test executed.

To cleanup our environment, we kill all processes with PIDs that we
find in those PID files before and after we execute the tests, just
in case if we for some reason did not exit cleanly and failed to
cleanup after ourselves.