Merge the two jest configurations into one
With the new capabilities of the TestEnvironment, we can simply
have one jest configuration and run tests in parallel.

This allows us to drop the `logDir` configuration of each test
suite. To still maintain nicely structured logs, we extract the
suite name from the testPath that we are passed in the constructor.