Download blockchain nodes in TS instead of with shell scripts
This has a number of advantages:

- The nodes are kept inside the `node_modules` folder and are
transparently downloaded if we need them on a new install.
This should also speed up CI because we are caching the node_modules
folder.
- It will ultimately allow us to remove a top-level directory from
the source tree, leaving behind a cleaner repository.
- We appropriately download those nodes into versioned folders.
- Typescript is easier to maintain compared to shell scripts.
- An implementation detail (the use of binaries instead of Docker
images) does not leak to the rest of the system.
- We will ultimately be able to simplify the API of the
BitcoindInstance and ParityInstance classes because we no longer
need to pass the project-root in.