Don't pass around the projectRoot variable
Resolving paths in different places is problematic because it
makes the system brittle. Instead, paths should only be resolved
in a central place and independently of each other.

With the blockchain nodes being cached in a specific location,
the only need for passing around the projectRoot is to resolve
the cargo target directory. Computing that from the projectRoot
is actually also prone to be wrong if we ever decide to change
where the e2e tests are placed. To be resilient against those
kind of changes, we use the `cargo metadata` subcommand to
know, where the target directory is.