Update parity to version 2.7.2 and start it with --force-direct
Parity has an auto-update mechanism where it actually downloads
the latest version and executes that binary instead of the current
one. It spawning a new process makes it impossible for us to kill
it with the PID we got from starting it.

The latest version (2.7.2 at this stage), seems to allow us to
circumvent this probem with the --force-direct flag.