Qualify `halight::State` usage
We just renamed `InvoiceState` to `State`.  It is now ambiguous _what_
State this refers to (not entirely true because the rfc003 version of
this object is still called `LedgerState` but when we either fix this or
add `han::State` then this statement will become true).

Qualify the state object with leading path `halight::State`.