Remove mod for traits
Trait are need to be imported in the file so having to specify
a module in the path does not have as much again than for structs.

Also, the name `event` is already used for the enum `Event`. Finally,
the trait does not represent an event per se but gives you the
functionality to know when an event has occurred.