Add a table for the walking skeleton created swap
We need a way to persist data for created swaps.  We would like to be
able to eventually restart swaps so we need everything that was handed
to us by the user via the REST API.  Also, we will at some stage need to
be able to link the data in 'created' swaps with the data for
'finalized' swaps, making a comment about this since it is currently out
of scope.