Move split protocol wrappers into swap_protocols
We have a wrapper for creating the swap for each of the protocols (han,
herc20, halight).  These are currently in `network/` but they are better
placed with the functions the wrap.  This has the added advantage of
making these functions private.  Thanks Lucas.

Move split protocol wrapper functions into their respective modules in
`swap_protocols/`.