Define the interface between controller and database
Users can create an ether/lightning swap via the REST API.  We would like a
way to save the data that the user provides into the database.  As a
first step in doing so define an interface between the controller and
the database.

This patch introduces a new naming convention for the split protocols.
Instead of fully describing them as we currently do, this patch proposes
the following convention

	HanEthereumEther -> HNETH    # Htlc (asset) Native ETH
	HalightLightningBitcoin -> LNBTC   # Lightning BTC

While not included in this patch this implies:

        HanBitcoinBitcoin -> HNBTC    # Htlc (asset) Native BTC

And assumes Herc20 will be shortened to:

	Herc20EthereumEther -> HERC20   # Htlc ERC-20

This naming convention allows modules to be placed under
`swap_protocols` for each concrete protocol i.e., `hnbtc`, `hneth`,
`lnbtc`, and `herc20`.  This allows for the shortening of the super long
identifiers we are actively adding e.g., `new_han_ethereum_ether_swap`
could become `hneth::new_swap`.