Remove the Lightning asset module
In hindsight it was a mistake to add the module `asset::lightning`.  By
definition the Lightning network does not have an associated asset,
Lightning can work on top of 'any' blockchain.  We should model it
accordingly.

Remove the `asset::Lightning` module.  Currently the only asset we
support on Lightning is bitcoin so use `asset::Bitcoin` instead.