Do not initiate a swap for identical parameters
If a swap has already been initiated we do not want to initiate a
second time otherwise there will be ambiguity in the system later during
the communication protocols since the swap digest is based off of the
created swap parameters.

There is more than one solution to this problem but for now just check
against the swap digest computed from the created swap parameters.

In order to be able to add a swap with the same digest again, once the
first has finalized, we need to remove the swap from the network
behaviour 'cache' i.e., all the hash maps that track the state as the
swap progresses through the communication protocols.

If a user attempts to post two identical swaps make the second attempt
return a 400 error.