Use debug instead of warn
Currently we have a warn message that lnd not supported for charlie,
this is expected behaviour however the warning message is quite
verbose (and brightly coloured in some terminals).  To make the e2e
tests more pleasant to run we can use a debug statement, this makes the
number of lines in the output lower and the colour more mute.