Move all swap hashmaps to a dedicated struct
Having several different hashmap for each swap information
and status creates complexity in `ComitLn` and also creates
the risk of having bugs that lead to discrepancy between
the different hashmaps.

By moving all hashmap to a given struct that has a consistent
interface, it allows the focus on avoiding discrepancy
to be localized to this new `Swaps` struct.

It also later allows the migration of this `Swaps` struct
to a state machine if we see it as a gain.