Separate relative and absolute time types
We currently use Timestamp and u32 as types for variables that represent
both relative and absolute times.

Absolute times are preferred but with the introduction of lightning,
relative times are unavoidable.

By ensuring that we use Timestamp (absolute) or RelativeTime whenever
We manipulate times and expiry, we are avoiding critical bugs.

The renaming of `Timestamp` to `AbsoluteTime` has been considered,
however, the `timestamp` terminology already carry a meaning of
 absoluteness.