Load protocol combination from database
To test the custom SQL function, we introduce proptest to the codebase.
Proptest is superiour to our existing property test framework "Quickcheck",
in the way that it is much easier to define custom strategies for
generating data because we don't have to implement a trait but can just
compose functions together.