Improve the network module
Builds on top of @D4nte's draft PR re-working the network module to make
it easier to add more protocols.

The idea is to make the Comit (previously named `ComitLN`)
NetworkBehaviour agnostic to the swap protocol that it is involved in
i.e., within Comit there is no alpha/beta only data to be communicated
with the peer node. One layer up, at the swarm (`ComitNode`
`NetworkBehaviour`) the protocol and role are known.

All reads and writes to the database, which imply knowledge of the
protocol, happen in the swarm. For this we add a reference to the new
Storage.