Do not mix event and data for Funded
`herc20::Funded`, described as "data available at given state"
actually was storing the fact on whether the htlc was correctly funded.
I believe this creates some confusion as we could see that the variable
returned by `wait_for_deployed` was actually named `event`. This trait
does not return the `enum Event`. The function trait resolving is the
event in itself, what is returned is the data.

In summary:
- (changed) `herc20::Funded` is a struct that only contains the data of
learned at a fund event
- (changed) When this data is returned, either `Event::Funded` or
`Event::IncorrectlyFunded` is emitted
- (same) Based on this event, the internal state is changed to `Funded`
or `IncorrectlyFunded`