Refactor make_siren_entity
We just added functionality to return a single action on the get swap
endpoint. This code is complex and a candidate for refactoring. Code
that is easy to read is easy to maintain and harbours less bugs.

Refactor `make_siren_entity` by splitting out helper functions and
simplifying the single action logic.

This patch is refactor only, no logic changes.