Use methods instead of trait DeriveIdentities
The abstraction provided by the `DeriveIdentities` is not needed, we
only ever do this for `SwapSeed`. This is an artefact of copying rfc003
code.

In order to keep rfc003 working 'as is', re-name
`rfc003::DeriveIdentities` to `rfc003::Rfc003DeriveIdentities`.

Define the two methods `derive_transient_refund_identity` and
`derive_transient_refund_identity` on `SwapSeed`.