Remove hard code ledger types
We have a bunch of places where we hard code the regtest network, for
all three ledgers. We already have this network information in the
database, all we need to do is expose it via the `Finalized` structs.

Add 'network' field, ether `ledger::Bitcoin` for Bitcoin and Lightning
swaps and `ChainId` for Ethereum, to the `Finalized` family of structs.

Use new fields and remove hard code ledger types.