Introduce within_swap_context macro
Used both in the HTTP API controller and when spawning swaps. The
design is exactly like it was for `rfc003_with_swap_types`, with some
type aliases not being used depending on the use-case.

Enabled by merging `http_api::SwapContext` and `spawn::SwapContext`
into one. There is no real reason to have two structs, even if we
sometimes ignore the swap ID because we already have it. Semantically,
the swap ID is part of the context of a swap.

Caveat: we have to implement `From<comit::Never for
ActionResponseBody` because the never type `!` is not stable yet (I
think). If it were stable, the compiler would know that certain action
implementations will only ever return an error (_never_ returning the
`Result::Ok` variant), so those non-existent actions don't need to
implement a conversion into `ActionResponseBody`.

Co-authored-by: Co-authored-by: Thomas Eizinger <thomas@eizinger.io>