http_api: Move type definitions
Code is easier to read if the most important stuff comes first. If
types exist solely for a function to use then the type expression is
less important than the function that uses it and can be placed below
the function.

Move the types used by `get_peers()` to below the function.