Implement LocalPeerId on the Facade
We currently get the local peer id from the `Rfc003Facade`. In order
to be able to move away from using the rfc003 facade we need this
functionality to be available when building the REST API routes.

Implement `LocalPeerId` on the `Facade` using delegate as is currently
done for the `Rfc003Facade`.