Define single hbit_herc20 function for Alice and Bob
The function can be used by both parties, ensuring that the execution
of a swap cannot get out of sync.

The function is generic over actors which have to implement certain
actions. In `WalletACTOR` implementations, a wallet is used to
actually peform the action; in `WatchOnlyACTOR` implementations,
`btsieve` is used to monitor the blockchain.

Temporary `BitcoinWallet` and `EthereumWallet` dummy types are
introduced to satisfy the compiler. They will eventually be replaced
by traits implemented by wallets (such as the ones which will actually
power Nectar).