Fix siren tests
Fix the siren tests to test that both the siren document is valid and
that the properties are valid according to our schema.

Remove the `swap.schema.json` because it does not do any real
validation, we can achieve this much by just validating the
"/swaps/:id" endpoint returns a valid siren document.