Improve the ergonomics of interacting with the swap_digest module
Instead of forcing our callers to instantiate a specific type and
call the `digest` function on it, we expose convenience functions
that allows us to compute a digest from anything that implements
the conversion function `Into`.

This also simplifies our import statements because our callers
only interact with the `swap_digest` module and can thereby remove
the direct dependency on the `Digest` crate.