Keep checking if we should abort when trying to do an action
Before, we only checked once, so we could end up stuck waiting for an
action that may never happen. We consider that we should stop waiting
for an action to take place as soon as the protocol "should be
aborted". For Nectar, this is as soon as Beta has expired.