Replace `if let Err` with `map_err`
While it achieves the same result, the usage of `if let` with an
execution call on the right handside can confused the reader such as
myself and make per believe that the call is conditional on returning
the error. I find `map_err` more readable as, to me, it is now obvious
that we are doing a function call and then processing the result.