Use forex trading terminology
The forex industry uses specific terminology. We are writing a
distributed orderbook for trading asset pairs - this is equivalent to
forex trading but in crypto. We should fit in with the world :)

Use quote and amount to construct an order, depends on the trading
pair (we have just BTC/DAI). Replaces the usage of two amounts as was
previously done.

For a swap of 10 BTC to 100_000 DAI (COMIT terminology) forex
terminology would be:

        tp: BTC/DAI
        quote: 10_000
        amount: 10

Whether this is a buy or sell order will determine which COMIT
protocol we use. Currently we assume order makers act in the role of
Bob.