Add support for partial order taking
The taker can specify a quantity in the take order request allowing the taker to partialy fill an order. Ethereum amount in the Order struct was replaced with a DAI to BTC conversion rate. The ethereum amount is calculated on the fly using this rate and the Bitcoin amount when a partial take request is made and received.