Add a dream API call for creating orders
This is purely fictional but this is how I'd like to call to cnd to
create a limit order. I assume I will later call cnd again supplying
identities to use when I confirm the take order request that would
come in if another node tried to take my order.