Separate refund from happy path in swap execution
This further simplifies the swap execution code, making it easier to
understand.

We do lose the ability to refund as Alice if Bob never redeems, but as
it is library code we don't need to be biased towards taking advantage
of the other party being inactive.

Also, move swap execution functions to specific `hbit_herc20` and
`herc20_hbit` modules.