Add config submodule to expiries
We model the swap state transitions by describing individual time
periods for actions that must occur in order to transition from one
state to the other.

Add a config module to provide the logic for calculating these
individual time periods based on pre-configured values.

Note: After much messing around with Poisson distributions it was
decided that simply using `average block time` * `n blocks` is just as
good a way to calculate the expected time to wait for a n blocks to be
mined as using the complex Poisson distribution calculation. This is
typical in the ecosystem. The biggest issue with this method is that
average block time is not known, this issue exists whether we use
Poisson distribution or not.