Replace import-alias with root crate path
Import aliases can be very confusing and surprising because their
names are usually made up on the spot and hence it is not clear
what exactly they refer to when reading the code.

Root crate paths (i.e. with a leading `::`) have an unambiguous meaning
and allows to import crates by name even if there is a module in scope
that shares the same name.

This allows us to stick to the familiar name of the crate without
having to invent new ones.