Save orders to the DB and allow them to be fetched again
There are several aspects about this implementation worth pointing out:

1. Co-location

We try and place most aspects around saving to and loading from the
DB within the single file that is concerned with it. This allows us
to make most of the structs and functions private to the module.
This is further improved by also having the warp filter definition
in the same module.

At this stage, this is an expirement in how we could structure things
differently in cnd.

2. DB design

The database design closely resembles what we have done for swaps
already. This is actually expected, given that we conceptually do the
same thing: Save parts of the data that we need for a swap to the
database. Due to our existing POST /swaps/.../... endpoints, a
refactoring is not possible at this stage. Once we get rid of these,
a cleaner design might be achievable.