Introduce `Price` and `Quantity` types
These newtypes allow us to contain all logic around calculations
of different rates and quote prices in a way that enforces the
correct unit.

The only way to construct instances of `Price` is through a "loud"
API that should make clear to the developer, what unit the price is
in. Currently, the only supported one here is WEI/SAT.