Remove Facade in favor of accessing Storage, Connectors & Swarm directly
The Facade is a relict from when we had a lot of traits for accessing
functionality from the controllers. This requirement no longer
exists and the facade is now a pretty useless layer that needs to
be maintained every time any of the function calls change.

We remove it in favor of passing Storage, Connectors and Swarm to
the controllers directly.