Introduce seed for network purposes
Instead of carrying around the root seed that is used to initialised
the crypto wallets, we used a `network::Seed` for network purposes.

The aim here is to protect the wallet seed from any bug in the network
module that would leak it.

It also removes the indirection of getting the seed via the Bitcoin
wallet.