Split LightningWallet into an LndClient and a LightningChannel
To allow for parallelization in the e2e tests AND accurate assertions,
we need to ensure each test gets its own channel. To make this
requirement easier to understand, we split the functionality into
a stateless LndClient class and a stateful LndChannel class.

An LndClient provides functionality to open a channel with another
client and returns an LndChannel upon success.

This LndChannel is equivalent to a wallet on other chains for our
tests: It has a balance and can perform actions like making sending
money (by making a payment).