Only fund an LndInstance once
This caused race conditions in CI. Conceptually, funding Lnd's wallet
during startup is also similar to how we treat the other ledgers.
They start with a single account that has a lot of money and can
distribute it to other accounts. Similarily, we start lnd with a lot
of money and distribute it to individual channels later for the actual
swap.