Allow user to specify desired swap protocol role through config
The role that nectar was previously hard coded to be bob. This PR adds an optional role field to the config file that allows the user to specify a role. If a role is not specified, Bob is chosen as a default to prevent a user unexpectedly losing fees from failed swaps.