Introduce FakeTreasuryService
Like other dependencies of the test environment, we start
said service only once for all tests.

At the moment, the rate is hardcoded to 9000 DAI per BTC.
If needed, we can make that configurable eventually.

We utilize the existing infrastructure of starting the ledger
nodes to start an HTTP server written with express. This
server implements a subset of the kraken ticker API. As such
we can point nectar against this service and thereby control
the rate at which nectar publishes orders.