Refactor release process to accomodate for nectar
We re-design the release process to use the new `workflow_dispatch`
event. This allows users to manually trigger a workflow with arbitrary
input. In our case, we use this to ask the user for the version and
then trigger the whole release process.

We also modularize the workflows into:

1. Triggering the release

This creates a release branch and opens a PR for it against master.

2. Creating a Github release

Once the release branch is merged, we create a Github release for
the given commit.

3. Building the release binaries

When a new release is created, we build the binaries for all platforms
and attach them to the release.