Close order after successful swap
After the hbit protocol completes successfully, we transition the
order state that is associated with the swap closed. This is a bit
hacky because ideally, we would want to do this at the end of swap.
However, cnd currently does not yet have a notion of a swap as it
only spawns two protocols. We can fix this once the comit lib exposes
a decent swap abstraction.