Remove fee subtraction from sell order creation
The max_sell represents the maximum amount to be traded for each
currency. This should not be concerned by the fees. Fees are on top
of the maximum amount.
Note that checks for sufficient funds are already in place, so we just
remove the fee subtraction, add unit tests to show the bahavior and
adap existing integration and e2e tests.