Remove ProtocolSpawner
ProtocolSpawner used to be a layer of indirection that simplified
spawning of protocols from various places.
However, the only place where it is called from was the `spawn` module.
Hence, we can remove the ProtocolSpawner in favor of doing whatever
we need to do within the spawn module.
Turns out, we don't even need a struct + trait abstraction for this.

This also improves the orthogonality of parameters because we are no
longer passing `Storage` twice to the `spawn` function. Prior to this
patch, we passed it explicitely once and implicitely through the
`ProtocolSpawner`.