Merge master into dev instead of the release branch
Previously, we've only merged the release branch back into dev after
a release. That ensured that any commits to the release branch like
bumping the version in the manifest and updating the changelog were
also present in the dev branch.

Even though those changes were merged back into dev, GitHub still
considered a new release branch to be "out-of-date" with master because
it was missing the _merge_ commit that represents the actual release.

There is no semantic difference between merging the release branch
and merging the master branch back into dev.

We change the workflow to use the master branch to have the warning
go away and avoid developers being distracted by it, thinking the
have to action it somehow.