Replace dedicated error structs with anyhow's .context
We don't need to define a struct for the error if we all we want
is to add more information to the error stacktrace. Addititionally,
if we were to define a struct for the error, we shouldn't print the
inner error as part of this error struct's message otherwise the
message will get printed twice once anyhow prints the error stacktrace.
More importantly, we want to include context that would otherwise
be lost like the URL we made the request to.