Do not use compare_and_swap for db update
Backward compatibility is maintained to allow newer software to
deserialize older data.
We do not try to guarantee that older data serialised the same way than
newer data.
Which means that `compare_and_swap` fails if the swap was created with
an older version of the software.

As we do not cater for concurrent update of the db, and we just retrieved
the swap to update few lines ago, it is ok to not check again.

This resolves the issue encountered when trying to archive a swap
created before the archive swap feature was introduced.