Only build the binaries instead of all targets for `make e2e`
Building all targets includes things like `cfg(test)` in the binaries.
However, the `CndInstance` and `NectarInstance` classes within the
e2e test framework will build the binaries with `cargo build -p`.
From the perspective of that command, the binary that was built using
the Makefile is outdated and needs to be re-built which leads to
unnecessarily long test execution times.